More renaming / moving of classes
